TcpServer can now handle responses. Module bugfix. SyncCall introduced.
The tcp server is now capable of receiving responses, in case the
server has sent a request to one of its clients.

Additionally the call can now also be executed synchronous and not
only asynchronous.

Added a new error "Not Yet Implemented", which is not required
by the library itself, but can be useful during development
of tool using this library.

Fixed a bug in the module, that caused an infinite recursion
on figuring out, if an id is required for a register callback.